[SwiftUI] Add clear background color to EpoxySwiftUIHostingController
A `UIHostingController` has a system background color by default as it's typically used in full-screen use cases. Since we're using this view controller to place SwiftUI views within other view controllers we default the background color to clear so we can see the views below, e.g. to draw highlight states in a `CollectionView`.