Update to Swift 5.4+
Swift 5.4 is included with Xcode 12.5, and lower versions of Swift/Xcode will no longer be able to submit to the App Store soon. As such we can bump the Swift version to Swift 5.4 in CocoaPods, SPM, and remove all of the conditional checks in the codebase.

While we're here we also update to the latest versions of the Swift packages we depend on.