[SwiftUI] Fix reused UIHostingController sizing
As of iOS 15.2, `UIHostingController` now renders updated content asynchronously, and as such this view will get sized incorrectly with the previous content when reused unless we invoke this semi-private API. We couldn't find any other method to get the view to resize synchronously after updating `rootView`, but hopefully this will become a public API soon so we can remove this call.

While we're here, and since SwiftUI has forced our hand to use semi-private APIs, we now migrate to using the `_disableSafeArea` API directly in the `EpoxySwiftUIHostingController`.