[SwiftUI] Support more sizing behaviors in measurement container
We want to additionally support a `UIView` that's intrinsically sized when hosted within a SwiftUI `View`, rather than expanding to fill its parent's offered bounds which is what the current behavior supports.

While we're here, we also rework the way that we compute sizes to no longer require a `DispatchQueue.main.async` by switching to a `ObservableObject` in the sizing context.