[SwiftUI] Rework SwiftUIMeasurementContainer to use sizeThatFits
This is the "correct" method to implement the behavior that we were achieving on iOS 13-15 using a `GeometryReader` as evidenced by the fact that Apple made this API public on iOS 16. As such, we move to it instead, and migrate off of the other approach. We're able to keep pretty much the same public facing API.