[SwiftUI] Rework SwiftUIMeasurementContainer to use sizeThatFits (#107)
* [SwiftUI] Rework SwiftUIMeasurementContainer to use sizeThatFits

This is the "correct" method to implement the behavior that we were achieving on iOS 13-15 using a `GeometryReader` as evidenced by the fact that Apple made this API public on iOS 16. As such, we move to it instead, and migrate off of the other approach. We're able to keep pretty much the same public facing API.

* Update CHANGELOG

* Remove extraneous UIViewType

* Add back logic to invalidate content size on subsequent bounds size changes

* Make sizing a chained setter rather than an additional parameter

* Rename cases to match SwiftUI naming

* Make measuredIntrinsicContentSize private, add clarification about differences